Refactors OpenVPN API calls
Moves OpenVPN server client API calls to a dedicated module.
This improves code organization and separation of concerns.
Also moves the data types to a types file.
