+++ src/pages/ServersOverview/index.tsx
@@ -2,8 +2,8 @@ import { useEffect, useMemo, useRef, useState } from "react";
 import { useParams } from "react-router-dom";
 import { toast } from "react-toastify";
 import DateRangeFilter, { type Grouping, type DateRangeChange } from "../../components/DateRangeFilter";
-import { fetchOverviewSeries, type OverviewSeriesResponse } from "../../utils/api";
-import { fetchOverviewTotals, type OverviewTotalsResponse } from "../../utils/api";
+import { fetchOverviewSeries, fetchOverviewTotals } from "../../utils/api/OpenVpnServerClients";
+import { type OverviewTotalsResponse, type OverviewSeriesResponse } from "../../utils/types";
 import StatsCards from "./StatsCards";
 import OverviewChart from "./OverviewChart";
 import GeoMap from "./GeoMap";
