+++ src/utils/api.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 import type { AxiosRequestConfig } from "axios";
-import type { Config, Certificate } from "../utils/types";
+import type { Config } from "../utils/types";
 import type { ApiResponse } from "../utils/api-response";
 import {
   HubConnectionBuilder,
@@ -132,154 +132,4 @@ export const getGeoLiteHubConnection = async (): Promise<HubConnection> => {
 export const logout = () => {
   localStorage.removeItem("token");
   window.location.href = "/login";
-};
-
-
-
-
-
-
-export type OverviewGrouping = "auto" | "hours" | "days" | "months" | "years";
-
-export type OverviewSeriesRow = {
-  ts: string;                 // ISO-8601 (UTC)
-  activeClients: number;
-  trafficInBytes: number;
-  trafficOutBytes: number;
-  trafficTotalBytes: number;
-};
-
-export type OverviewSeriesMeta = {
-  from: string;               // ISO-8601
-  to: string;                 // ISO-8601
-  grouping: Exclude<OverviewGrouping, "auto">;
-  timezone: string;
-  trafficUnit: "bytes";       // <-- like on backend
-  vpnServerId?: number | null;
-};
-
-export type OverviewSeriesSummary = {
-  totalTrafficInBytes: number;
-  totalTrafficOutBytes: number;
-  peakActiveClients: number;
-};
-
-export type OverviewSeriesResponse = {
-  meta: OverviewSeriesMeta;
-  summary: OverviewSeriesSummary;
-  series: OverviewSeriesRow[];
-};
-
-export type FetchOverviewSeriesParams = {
-  from: Date;
-  to: Date;
-  grouping?: OverviewGrouping;   // default "auto"
-  vpnServerId?: number | null;
-  externalId?: string | null;
-};
-
-// always serialize as UTC ISO (…Z)
-const toUtcIso = (d: Date | string | number) => {
-  const dt = d instanceof Date ? d : new Date(d);
-  return dt.toISOString();
-};
-
-export const fetchOverviewSeries = async (
-  params: FetchOverviewSeriesParams
-): Promise<OverviewSeriesResponse> => {
-  const { from, to, grouping = "auto", vpnServerId, externalId } = params;
-
-  const qs = new URLSearchParams();
-  qs.set("from", toUtcIso(from));
-  qs.set("to", toUtcIso(to));
-  qs.set("grouping", grouping);
-  if (vpnServerId != null) qs.set("vpnServerId", String(vpnServerId));
-  if (externalId && externalId.trim()) qs.set("externalId", externalId.trim());
-
-  const res = await apiRequest<OverviewSeriesResponse>(
-    "get",
-    `/OpenVpnServerClients/overview/series?${qs.toString()}`
-  );
-  return res.data;
-};
-
-export const fetchGeoPoints = async (
-  params: FetchGeoPointsParams
-): Promise<GeoPointAggDto[]> => {
-  const { from, to, vpnServerId, externalId, onlyWithCoordinates = true } = params;
-
-  const qs = new URLSearchParams();
-  qs.set("from", toUtcIso(from));
-  qs.set("to", toUtcIso(to));
-  qs.set("onlyWithCoordinates", String(onlyWithCoordinates));
-  if (vpnServerId != null) qs.set("vpnServerId", String(vpnServerId));
-  if (externalId && externalId.trim()) qs.set("externalId", externalId.trim());
-
-  const res = await apiRequest<GeoPointAggDto[]>(
-    "get",
-    `/OpenVpnServerClients/overview/points?${qs.toString()}`
-  );
-  return res.data;
-};
-
-export const fetchOverviewTotals = async (
-  params: FetchOverviewTotalsParams
-): Promise<OverviewTotalsResponse> => {
-  const { from, to, vpnServerId, externalId } = params;
-
-  const qs = new URLSearchParams();
-  qs.set("from", toUtcIso(from));
-  qs.set("to", toUtcIso(to));
-  if (vpnServerId != null) qs.set("vpnServerId", String(vpnServerId));
-  if (externalId && externalId.trim()) qs.set("externalId", externalId.trim());
-
-  const res = await apiRequest<OverviewTotalsResponse>(
-    "get",
-    `/OpenVpnServerClients/overview/summary?${qs.toString()}`
-  );
-  return res.data;
-};
-
-
-export type GeoPointAggDto = {
-  country: string | null;
-  region: string | null;
-  latitude: number | null;
-  longitude: number | null;
-  sessionsCount: number;
-  totalBytesIn: number;
-  totalBytesOut: number;
-};
-
-export type FetchGeoPointsParams = {
-  from: Date | string;
-  to: Date | string;
-  vpnServerId?: number | null;
-  externalId?: string | null;
-  onlyWithCoordinates?: boolean;
-};
-
-export type OverviewTotalsResponse = {
-  meta: {
-    from: string;        // ISO
-    to: string;          // ISO
-    grouping: string;    // "none"
-    timezone: string;    // "UTC"
-    trafficUnit: string; // "bytes"
-    vpnServerId?: number | null;
-  };
-  totals: {
-    sessionsCount: number;
-    usersCount: number;          // unique ExternalId
-    trafficInBytes: number;
-    trafficOutBytes: number;
-    trafficTotalBytes: number;
-  };
-};
-
-export type FetchOverviewTotalsParams = {
-  from: Date | string;
-  to: Date | string;
-  vpnServerId?: number | null;
-  externalId?: string | null;
 };
